Create Connecting map by truck trips at Laemchabung port Thailand
This is the part of the research method about logistics project. The content is study of cost shipping by cargo ships at Laemchabung port Thailand. 
For the part of coding is retrieving data from GPS that are installed in each trucks in order to calcultate proportion of truck traveling from origin to destination in each province in Thailand.